+++ dev/src/DataAccessLayer/DalObjects/InventoryObjects/DalItem.java
@@ -29,8 +29,7 @@ public class DalItem extends DalObject<DalItem> {
     private int storageQuantity;
     private String shelfLocation;
     private String storageLocation;
-
-    private double weight;
+    private int weight;
     private String categoryName;
 
     public DalItem() throws SQLException {
@@ -139,11 +138,11 @@ public class DalItem extends DalObject<DalItem> {
         this.storageLocation = storageLocation;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
 
-    public void setWeight(double weight) {
+    public void setWeight(int weight) {
         this.weight = weight;
     }
