+++ dev/src/ServiceLayer/InventoryService.java
@@ -17,7 +17,7 @@ public interface InventoryService {
 
     Response addItem(int id, String name, String categoryName, double costPrice, double sellingPrice, int minAmount,
                      String shelfLocation, String storageLocation, int shelfQuantity, int storageQuantity, int manufacturerId,
-                     double weight, List<String> suppliersIds);
+                     int weight, List<String> suppliersIds);
     ResponseT<FacadeItem> getItem(int itemId);
     Response modifyItemName(int itemId, String newName);
     Response modifyItemCategory(int itemId, String newCategoryName);
