+++ dev/src/ServiceLayer/Service.java
@@ -311,7 +311,7 @@ public class Service implements IService {
     }
 
     @Override
-    public Response addItem(int id, String name, String categoryName, double costPrice, double sellingPrice, int minAmount, String shelfLocation, String storageLocation, int shelfQuantity, int storageQuantity, int manufacturerId, double weight, List<String> suppliersIds) {
+    public Response addItem(int id, String name, String categoryName, double costPrice, double sellingPrice, int minAmount, String shelfLocation, String storageLocation, int shelfQuantity, int storageQuantity, int manufacturerId, int weight, List<String> suppliersIds) {
         return inventoryService.addItem(id,name,categoryName,costPrice,sellingPrice,minAmount,shelfLocation,storageLocation,shelfQuantity,storageQuantity,manufacturerId,weight,suppliersIds);
     }
